backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
dnspython==2.8.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unicorn==2.1.4
urllib3==2.5.0
uvicorn==0.37.0
